+++ CHANGELOG.md
@@ -6,11 +6,12 @@
 * Added Content feature methods:
   * `enterContentZone`, to start Content checks (Experimental!)
   * `exitContentZone`, to stop Content checks (Experimental!)
+* Added `Countly.deviceId.setID` method for changing device ID based on the device ID type
 * Mitigated an issue where a session could have started while the app was in the background when the device ID was changed (non-merge).
 * Deprecated following SDK calls:
   * `Countly.getCurrentDeviceId` (replaced with: `Countly.deviceId.getID`)
   * `Countly.getDeviceIDType` (replaced with: `Countly.deviceId.getType`)
-  * `Countly.changeDeviceId`
+  * `Countly.changeDeviceId` (replaced with: `Countly.deviceId.setID`)
 * Android Specific Changes:
   * ! Minor breaking change ! Unsupported types for user properties will now be omitted, they won't be converted to strings.
   * Disabled caching for webviews.